requirements.txt
opencv-python==4.5.5.64
Pillow==9.1.0
numpy==1.22.3
matplotlib==3.5.2


